+++ site/dat/docs/Changelog.md
@@ -6,6 +6,7 @@
  - fix public cloud reporting (unauthorised notes sending)
  - add test for paths of resource files 
  - bump up auto-installed JMeter to 3.1 and pmgr to 0.11
+ - add limitation of response time list size to cumulative KPISets
 
 ## 1.7.4 <sup>11 nov 2016</sup>
  - fix Locust crash when used with 'requests'-style scenario and cloud provisioning